Adb Push/pull;  Clarify pathname; rename magisk_patched.img
Instead of "Copy recovery image to device" (Your note that MTP can corrupt files is too far down in the document)
  ,,, Use adb to push the boot/recovery image from your pc to your device. This method preserves the file from corruption, which may happen if you copy and paste, or drag and drop from your pc to the device-directory path. 
                    ADB push <filepath>/recovery.img /sdcard/Download

As these simple instructions are for the brief introduction to adb.  In push/pull commands, the pathnames used for the sdcard needs some explanation.  ( I had to learn the hard way due to the lack of explanation in these instructions)<br>
               (Don't be confused by your PC's linked-names for your android device storage. Windows' [Internal Shared Storage] directory is the /sdcard directory on your android device)

My phone did not recognize the "magisk_patched.img"  as named as a recovery file.  I had to rename it to "recovery.img". Your instructions fail to tell to rename the file before flashing it to the phone.